Feat/54 : 웹툰 생성 api 추가 (#59)
* chore: Webflux 의존성 추가

* feat : AiServer Property추가

* feat : WebClientConfig 추가

* feat : AiAuthorRepository & NewsSourceRepository 추가

* chore : .env.example 추가

* feat : webtoon 생성 api 추가 및 테스트 api 추가

* feat : redis 설정 추가

+++ src/main/java/com/akatsuki/newsum/domain/webtoon/dto/WebtoonSlideDto.java
@@ -1,7 +1,7 @@
 package com.akatsuki.newsum.domain.webtoon.dto;
 
 public record WebtoonSlideDto(
-	int slideSeq,
+	Byte slideSeq,
 	String imageUrl,
 	String content
 ) {
